// pc/src/app/types/common.type.ts
export interface ResponseBody {
    data: any;
    msg: string;
    status: number;
}
// ...
export interface GameList {
    gameCode: string;
    gameName: string;
    gameImg: string;
    balance: any;
}
// ...
export interface UserInfo {
    username: string;
    vipLevel: string;
    wallet: number;
    email: string;
    hasBankCard: boolean;
    hasMobile: boolean;
    hasWithdrawPassword: boolean;
    integral: number;
    loginDate: string;
    regDate: string;
    mobile: string;
    realname: string;
    rmk: string;
    totalBalance: number;
    uid: number;
    unread: number;
    qq: string;
    weixin: string;
    agencyLevel: string;              // 代理会员层级
    withdraw: boolean;                // 是否可以提取佣金
    agencyStatus: number;             // 代理状态    0-加入 1-停用 2-未加入
    outstandingCommissions: string;   // 未结佣金
    commissionBeginDate: string;      // 计佣开始时间
    commissionEndDate: string;        // 计佣结束时间
}